Small note component move
